Move stdlib-specific reachability test to LibraryReachabilityTest.
These tests are not testing the Analyzer, but rather the
arrangement of the standard library.
